comps/retrievers/neo4j/llama-index/requirements.txt
docarray[full]
fastapi
frontend
huggingface_hub
llama_index_graph_stores_neo4j
llama-index-core
llama-index-llms-openai
llama-index-llms-text-generation-inference
llama-index-embeddings-text-embeddings-inference
neo4j
numpy
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-sdk
pandas
Pillow
prometheus-fastapi-instrumentator
pydantic
pymupdf
python-docx
sentence_transformers
shortuuid
tiktoken
uvicorn
